Moved the base class manager logic to the clFork package.  Added empty test for multi inheritance testing
